+++ integration/frameworks/vue/vue-procaptcha-integration-demo/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @prosopo/vue-procaptcha-integration-demo
 
+## 1.1.88
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.88
+
+## 1.1.87
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.87
+
+## 1.1.86
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.86
+
+## 1.1.85
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.85
+
+## 1.1.84
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.84
+
+## 1.1.83
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.83
+
+## 1.1.82
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.82
+
+## 1.1.81
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.81
+
+## 1.1.80
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.80
+
+## 1.1.79
+### Patch Changes
+
+  - @prosopo/vue-procaptcha-wrapper@1.1.79
+
 ## 1.1.78
 ### Patch Changes